Split hardware clients.
